fix: compile errors after moving classes
